requirements.txt
PyQt6>=6.6.0
qtawesome>=1.3.0
pycryptodome>=3.19.0
argon2-cffi>=23.1.0
Pillow>=10.2.0
piexif>=1.1.3
psutil>=5.9.0
PySocks>=1.7.1
paramiko>=3.4.0
packaging